+++ senlin/objects/requests/nodes.py
@@ -10,6 +10,8 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
+from oslo_utils import versionutils
+
 from senlin.common import consts
 from senlin.objects import base
 from senlin.objects import fields
@@ -75,10 +77,25 @@ class NodeUpdateRequest(base.SenlinObject):
 
 @base.SenlinObjectRegistry.register
 class NodeDeleteRequest(base.SenlinObject):
+    # VERSION 1.0: Initial version
+    # VERSION 1.1 Added field 'force'
+    VERSION = '1.1'
+    VERSION_MAP = {
+        '1.8': '1.1',
+    }
 
     fields = {
-        'identity': fields.StringField()
+        'identity': fields.StringField(),
+        'force': fields.BooleanField(default=False)
     }
+
+    def obj_make_compatible(self, primitive, target_version):
+        super(NodeDeleteRequest, self).obj_make_compatible(
+            primitive, target_version)
+        target_version = versionutils.convert_version_to_tuple(target_version)
+        if target_version < (1, 1):
+            if 'force' in primitive['senlin_object.data']:
+                del primitive['senlin_object.data']['force']
 
 
 @base.SenlinObjectRegistry.register