+++ senlin/tests/unit/engine/service/test_clusters.py
@@ -1890,7 +1890,7 @@ class ClusterTest(base.SenlinTestCase):
         mock_policies.return_value = []
         mock_receivers.return_value = []
         mock_action.return_value = 'ACTION_ID'
-        req = orco.ClusterDeleteRequest(identity='IDENTITY')
+        req = orco.ClusterDeleteRequest(identity='IDENTITY', force=False)
 
         result = self.eng.cluster_delete(self.ctx, req.obj_to_primitive())
 
@@ -1913,7 +1913,8 @@ class ClusterTest(base.SenlinTestCase):
         cluster = mock.Mock(id='cluster1', status='ACTIVE',
                             dependents=dependents)
         mock_find.return_value = cluster
-        req = orco.ClusterDeleteRequest(identity='FAKE_CLUSTER')
+        req = orco.ClusterDeleteRequest(identity='FAKE_CLUSTER',
+                                        force=False)
 
         ex = self.assertRaises(rpc.ExpectedException,
                                self.eng.cluster_delete,
@@ -1928,7 +1929,8 @@ class ClusterTest(base.SenlinTestCase):
     def test_cluster_delete_not_found(self, mock_find):
         mock_find.side_effect = exc.ResourceNotFound(type='cluster',
                                                      id='Bogus')
-        req = orco.ClusterDeleteRequest(identity='Bogus')
+        req = orco.ClusterDeleteRequest(identity='Bogus',
+                                        force=False)
 
         ex = self.assertRaises(rpc.ExpectedException,
                                self.eng.cluster_delete,
@@ -1944,7 +1946,8 @@ class ClusterTest(base.SenlinTestCase):
                            consts.CS_DELETING, consts.CS_RECOVERING]:
             fake_cluster = mock.Mock(id='12345678AB', status=bad_status)
             mock_find.return_value = fake_cluster
-            req = orco.ClusterDeleteRequest(identity='BUSY')
+            req = orco.ClusterDeleteRequest(identity='BUSY',
+                                            force=False)
 
             ex = self.assertRaises(rpc.ExpectedException,
                                    self.eng.cluster_delete,
@@ -1961,7 +1964,8 @@ class ClusterTest(base.SenlinTestCase):
         x_obj = mock.Mock(id='12345678AB', dependents={})
         mock_find.return_value = x_obj
         mock_policies.return_value = [mock.Mock()]
-        req = orco.ClusterDeleteRequest(identity='IDENTITY')
+        req = orco.ClusterDeleteRequest(identity='IDENTITY',
+                                        force=False)
 
         ex = self.assertRaises(rpc.ExpectedException,
                                self.eng.cluster_delete,
@@ -1983,7 +1987,8 @@ class ClusterTest(base.SenlinTestCase):
         mock_find.return_value = x_obj
         mock_policies.return_value = []
         mock_receivers.return_value = [mock.Mock()]
-        req = orco.ClusterDeleteRequest(identity='IDENTITY')
+        req = orco.ClusterDeleteRequest(identity='IDENTITY',
+                                        force=False)
 
         ex = self.assertRaises(rpc.ExpectedException,
                                self.eng.cluster_delete,
@@ -2007,7 +2012,8 @@ class ClusterTest(base.SenlinTestCase):
         mock_find.return_value = x_obj
         mock_policies.return_value = [mock.Mock()]
         mock_receivers.return_value = [mock.Mock()]
-        req = orco.ClusterDeleteRequest(identity='IDENTITY')
+        req = orco.ClusterDeleteRequest(identity='IDENTITY',
+                                        force=False)
 
         ex = self.assertRaises(rpc.ExpectedException,
                                self.eng.cluster_delete,
@@ -2022,3 +2028,35 @@ class ClusterTest(base.SenlinTestCase):
         mock_policies.assert_called_once_with(self.ctx, '12345678AB')
         mock_receivers.assert_called_once_with(
             self.ctx, filters={'cluster_id': '12345678AB'})
+
+    @mock.patch.object(am.Action, 'create')
+    @mock.patch.object(ro.Receiver, 'get_all')
+    @mock.patch.object(cpo.ClusterPolicy, 'get_all')
+    @mock.patch.object(co.Cluster, 'find')
+    @mock.patch.object(dispatcher, 'start_action')
+    def test_cluster_delete_force(self, notify, mock_find, mock_policies,
+                                  mock_receivers, mock_action):
+        for bad_status in [consts.CS_CREATING, consts.CS_UPDATING,
+                           consts.CS_DELETING, consts.CS_RECOVERING]:
+            x_obj = mock.Mock(id='12345678AB', status=bad_status,
+                              dependents={})
+            mock_find.return_value = x_obj
+            mock_policies.return_value = []
+            mock_receivers.return_value = []
+            mock_action.return_value = 'ACTION_ID'
+            req = orco.ClusterDeleteRequest(identity='IDENTITY', force=True)
+
+            result = self.eng.cluster_delete(self.ctx, req.obj_to_primitive())
+
+            self.assertEqual({'action': 'ACTION_ID'}, result)
+            mock_find.assert_called_with(self.ctx, 'IDENTITY')
+            mock_policies.assert_called_with(self.ctx, '12345678AB')
+            mock_receivers.assert_called_with(
+                self.ctx, filters={'cluster_id': '12345678AB'})
+            mock_action.assert_called_with(
+                self.ctx, '12345678AB', 'CLUSTER_DELETE',
+                name='cluster_delete_12345678',
+                cause=consts.CAUSE_RPC,
+                status=am.Action.READY)
+
+            notify.assert_called_with()